refactor: Seller 도메인을 domain.seller 패키지 하위로 옮긴다
